Projectiles now kill enemies!
The projectiles now kill enemies! So far, they can only one-shot the
enemies, because the projectiles do 20 damage to the enemies, and the
enemies have 40 life. This sounds good, but the enemies take damage when
the projectile is NEAR them, which can occur in 2 consecutive frames, so
it appears that they one-shot the enemies. We will fix this by giving
the enemies invulnerability after they take damage (we need to do this
with players too.)