I added AI
There is now enemy AI, in 4 phases. I explain the phases within comments
of the code. However, phase 4 (the random direction) AI is not working
as intended. Some should try to fix that. It goes in a different
direction each frame, when it SHOULD pick 1 direction, and then go in
that direction for 300 frames.